Project 3
This programming assignment has three main learning goals. First, the assignment will provide you with an opportunity to practice your skills in developing programs in JavaTM using the Eclipse integrated development environment. Second, this assignment will allow you to practice imple- menting the decision theory principle of maximum expected utility in a game-playing context. Your work on this assignment is intended to strengthen your understanding of how expected utility values are calculated, including how they are calculated in multi-agent environments. Finally, this assignment requires you to design a heuristic evaluation function for a game, providing you with experience in the design of such functions. 
You will be provided with JavaTM source code that implements a two-player version of the Zombie DiceTM game, with a single computer player pitted against a single human player. This provided code will be fully functional, but the computer player performs very poorly in this version, due to missing components in its move selection algorithm. You will be required to complete the implementation of the computer player’s decision procedure by writing a function that calculates an expected utility value and another that computes a heuristic evaluation value for a state of play.