Project 4
This programming assignment has two main learning goals. First, the assignment will provide you with an opportunity to practice your skills in developing programs in JavaTM using the Eclipse integrated development environment. Second, this assignment will provide you with some experi- ence in implementing a key component of the most widely used artificial neural network learning procedure, namely the backpropagation of error learning algorithm. Your efforts in implementing this procedure are intended to produce a deeper understanding of machine learning methods using artificial neural networks. At minimum, completing this assignment will require an understand- ing of how error information can be propagated backward in a multi-layer perceptron in order to support the modification of connection weights in a manner that improves the performance of the network. 
You will be provided with JavaTM source code that implements a basic version of the back- propagation algorithm. This code will be complete except for the methods that calculate the delta (error) values for both output processing units and hidden processing units. You will be required to complete this implementation by supplying these methods with bodies, allowing the program to appropriately learn from a set of input-output training patterns.